+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 0.1.3
 - `WebSockex.start_link` will no longer cause the calling process to exit on
   connection failure and will return a proper error tuple instead.
 - Change `WebSockex.Conn.RequestError` to `WebSockex.RequestError`.